Cleanup and add comments
